+++ src/main/java/com/ewok/twitchmeme/dto/post/PostSaveRequestDto.java
@@ -1,5 +1,6 @@
-package com.ewok.twitchmeme.dto;
+package com.ewok.twitchmeme.dto.post;
 
+import com.ewok.twitchmeme.domain.member.Member;
 import com.ewok.twitchmeme.domain.post.Post;
 import com.ewok.twitchmeme.domain.post.Youtube;
 import lombok.Builder;
@@ -8,20 +9,23 @@ import lombok.NoArgsConstructor;
 
 import java.util.ArrayList;
 
-@NoArgsConstructor
 @Getter
-public class PostUpdateRequestDto {
+@NoArgsConstructor
+public class PostSaveRequestDto {
 
+    private Long memberId;
     private String title;
     private String summary;
+    private String broadcastId;
     private String content;
     private ArrayList<String> reference;
 
-
     @Builder
-    public PostUpdateRequestDto(String title, String summary, String content, ArrayList<String> reference) {
+    public PostSaveRequestDto(Long memberId, String title, String summary, String broadcastId, String content, ArrayList<String> reference) {
+        this.memberId = memberId;
         this.title = title;
         this.summary = summary;
+        this.broadcastId = broadcastId;
         this.content = content;
         this.reference = reference;
     }
