+++ src/main/java/com/ewok/twitchmeme/dto/post/PostUpdateRequestDto.java
@@ -1,6 +1,5 @@
-package com.ewok.twitchmeme.dto;
+package com.ewok.twitchmeme.dto.post;
 
-import com.ewok.twitchmeme.domain.member.Member;
 import com.ewok.twitchmeme.domain.post.Post;
 import com.ewok.twitchmeme.domain.post.Youtube;
 import lombok.Builder;
@@ -9,23 +8,20 @@ import lombok.NoArgsConstructor;
 
 import java.util.ArrayList;
 
-@Getter
 @NoArgsConstructor
-public class PostSaveRequestDto {
+@Getter
+public class PostUpdateRequestDto {
 
-    private Long memberId;
     private String title;
     private String summary;
-    private String broadcastId;
     private String content;
     private ArrayList<String> reference;
 
+
     @Builder
-    public PostSaveRequestDto(Long memberId, String title, String summary, String broadcastId, String content, ArrayList<String> reference) {
-        this.memberId = memberId;
+    public PostUpdateRequestDto(String title, String summary, String content, ArrayList<String> reference) {
         this.title = title;
         this.summary = summary;
-        this.broadcastId = broadcastId;
         this.content = content;
         this.reference = reference;
     }
