+++ src/main/java/com/ewok/twitchmeme/service/PostService.java
@@ -3,18 +3,14 @@ package com.ewok.twitchmeme.service;
 import com.ewok.twitchmeme.domain.member.Member;
 import com.ewok.twitchmeme.domain.member.MemberRepository;
 import com.ewok.twitchmeme.domain.post.*;
-import com.ewok.twitchmeme.dto.PostResponseDto;
-import com.ewok.twitchmeme.dto.PostSaveRequestDto;
-import com.ewok.twitchmeme.dto.PostUpdateRequestDto;
-import com.ewok.twitchmeme.dto.PostsDetailResponseDto;
+import com.ewok.twitchmeme.dto.post.PostResponseDto;
+import com.ewok.twitchmeme.dto.post.PostSaveRequestDto;
+import com.ewok.twitchmeme.dto.post.PostUpdateRequestDto;
+import com.ewok.twitchmeme.dto.post.PostsDetailResponseDto;
 import com.ewok.twitchmeme.dto.post.PostPagingListResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
