+++ src/main/java/com/ewok/twitchmeme/service/ReplyService.java
@@ -6,8 +6,8 @@ import com.ewok.twitchmeme.domain.post.Post;
 import com.ewok.twitchmeme.domain.post.PostRepository;
 import com.ewok.twitchmeme.domain.post.Reply;
 import com.ewok.twitchmeme.domain.post.ReplyRepository;
-import com.ewok.twitchmeme.dto.ReplyRequestDto;
-import com.ewok.twitchmeme.dto.ReplyResponseDto;
+import com.ewok.twitchmeme.dto.reply.ReplyRequestDto;
+import com.ewok.twitchmeme.dto.reply.ReplyResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
